[Ray] `DocumentFilter` and `Filter`/`Score`/`ScoreFilter` (#746)
* add documentfilter implementation


* fix nits and ruff


* add additional logic for setup, setup_on_node, and process_batch


* add pytests


* add dep


* more dep edits


* another dep


* add fasttext dep


* add jieba and mecab


* add default None params for setup_on_node and setup functions


* add praateek's suggestions


* organize imports


* remove process_batch


* add _metadata to result


* add praateek's suggestions


* ruff and post init for _name


* modify test


---------
